mega update

+++ index.js
@@ -6,7 +6,11 @@ const client = new Discord.Client();
 const d = new Date()
 const messages = []
 const report = []
+const banns = []
+const kicks = []
+const adlog = []
 const botName = config["bot-name"]
+const adminToken = config.adminToken
 client.on('ready', () => {
     for(let i=0;i<10;i++){console.log()}
     console.log("You're running Bot version: ["+config.version+"]")
@@ -16,25 +20,10 @@ client.on('ready', () => {
 client.on('message', msg => {
   let currentTime = d.getTime()
   messages.push(`AUTHOR: ${msg.author.username} | MESSAGE: ${msg.content} | TIME: ${currentTime}`)
-  //COMMANDS
-  if (msg.content === '!ping') {
-    msg.send('Pong!')
-  }
+  
   //RANDOM EMOJI COMMAND
   if (msg.content === '!emoji'){
       msg.reply(emoji[Math.floor(Math.random()*emoji.length)])
-  }
-  //CONSOLE LOG: ALL MESSAGES
-  if (msg.content === 'messages'){
-    if(!msg.member.hasPermission('EDIT_MESSAGES')){
-      msg.channel.send('Na tento prikaz nemas opravneni! { '+msg.member+" }")
-      msg.delete()
-      return
-    }
-    msg.channel.send('ZPRAVY')
-      for(let i=0;i<messages.length;i++){
-        msg.channel.send(messages[i])
-      }
   }
 });
 
@@ -43,15 +32,42 @@ client.on('message', message => {
   if (!message.guild) return;
   //KICK
   if (message.content.startsWith('!kick')) {
+    if(message.channel.id != "646099774797250575"){
+      message.delete()
+      return
+    }
+    if(!message.member._roles.includes(adminToken)){
+      message.reply("Na tohle nemas prava")
+      return
+    }
+    const args = message.content.split(' ').slice(1);
+    const kickReason = args.slice(1).join(' '); 
+    if(!kickReason){
+      message.reply("Prosim pouzij /kick <hrac> <duvod>")
+      return
+    }
     const user = message.mentions.users.first();
     if (user) {
       const member = message.guild.member(user);
       if (member) {
+        kicks.push(user.username+"#"+user.discriminator)
+        kicks.push(kickReason)
+        adlog.push(message.author.username+"#"+message.author.discriminator+" kicked "+user.username+"#"+user.discriminator)
         member
           .kick('Optional reason that will display in the audit logs')
           .then(() => {
-            message.reply(`Successfully kicked ${user.tag}`);
-          })
+            const exampleEmbed = new Discord.MessageEmbed()
+            .setColor('#0099ff')
+            .setTitle('Kick')
+            .setDescription("Hrac byl uspense kicknut")
+            .addFields(
+              { name: "**Hrac**",value: user.username+"#"+user.discriminator},
+              { name: "**Duvod**", value: banReason}
+            )
+            .setTimestamp()
+          
+            message.channel.send(exampleEmbed);
+                    })
           .catch(err => {
             message.reply('I was unable to kick the member');
             console.error(err);
@@ -65,16 +81,44 @@ client.on('message', message => {
   }
   //BAN
   if (message.content.startsWith('!ban')) {
+    //CHECK IF USER HAVE ROLE
+    if(message.channel.id != "646099774797250575"){
+      message.delete()
+      return
+    }
+    if(!message.member._roles.includes(adminToken)){
+      message.reply("Na tohle nemas prava")
+      return
+    }
+    const args = message.content.split(' ').slice(1);
+    const banReason = args.slice(1).join(' '); 
+    if(!banReason){
+      message.reply("Prosim pouzij /ban <hrac> <duvod>")
+      return
+    }
     const user = message.mentions.users.first();
     if (user) {
       const member = message.guild.member(user);
       if (member) {
+        adlog.push(message.author.username+"#"+message.author.discriminator+" BANNED "+user.username+"#"+user.discriminator)
+        banns.push(user.username+"#"+user.discriminator)
+        banns.push(banReason)
         member
           .ban({
             reason: 'They were bad!',
           })
           .then(() => {
-            message.reply(`Successfully banned ${user.tag}`);
+            const exampleEmbed = new Discord.MessageEmbed()
+	.setColor('#0099ff')
+	.setTitle('Ban')
+	.setDescription("Hrac byl uspense zabanovan")
+	.addFields(
+    { name: "**Hrac**",value: user.username+"#"+user.discriminator},
+    { name: "**Duvod**", value: banReason}
+	)
+	.setTimestamp()
+
+  message.channel.send(exampleEmbed);
           })
           .catch(err => {
             message.reply('I was unable to ban the member');
@@ -89,26 +133,31 @@ client.on('message', message => {
   }
   //REPORT
   if (message.content.startsWith('!report')) {
+    const args = message.content.split(' ').slice(1);
+    const reportReason = args.slice(1).join(' '); 
+    if(!reportReason){
+      message.reply("Prosim pouzij /report <hrac> <duvod>")
+      return
+    }
+
     const user = message.mentions.users.first();
     if (user) {
       const member = message.guild.member(user);
       if (member) {
         message.delete()
-        console.log("REPORTED: ["+user.username+" #"+user.discriminator+"]")
-        report.push(user.username+"#"+user.discriminator)
+        console.log("REPORTED: ["+user.username+"#"+user.discriminator+" for:"+reportReason+"]")
+        report.push(user.username +"#"+user.discriminator)
+        report.push(reportReason)
 
         const exampleEmbed = new Discord.MessageEmbed()
 	.setColor('#0099ff')
 	.setTitle('Report')
-	.setDescription("Uzivatel ["+ user.username+"#"+user.discriminator +"] byl nahlasen")
+	.setDescription("Hrac byl uspense nahlasen")
 	.addFields(
-		{ name:"Dekujeme",value: 'Admin co nejrychleji vyresi tvuj problem' },
-		// { name: 'Inline field title', value: 'Some value here', inline: true },
+    { name: "**Hrac**",value: user.username+"#"+user.discriminator},
+    { name: "**Duvod**", value: reportReason}
 	)
-	// .addField('Inline field title', 'Some value here', true)
-	// .setImage('https://i.imgur.com/wSTFkRM.png')
 	.setTimestamp()
-	// .setFooter('Some footer text here', 'https://i.imgur.com/wSTFkRM.png');
 
   message.channel.send(exampleEmbed);
 
@@ -120,24 +169,79 @@ client.on('message', message => {
     }
   }
   if(message.content.startsWith('!list')) {
+    if(!message.member._roles.includes(adminToken)){
+      message.reply("Na tohle nemas prava")
+      return
+    }
+
+
     if(report.length == 0){
       message.channel.send("Zadny hrad nebyl reportovan")
       return
     }
     const exampleEmbed = new Discord.MessageEmbed()
     .setColor('#0099ff')
-    .setTitle('Report List')
+    .setTitle('**Report List**')
     .setThumbnail()
     .addFields(
-        {name:"Reportovani hraci", value: report } 
+        {name:"**Reportovani hraci**", value: report} 
     )
     .setTimestamp()
   
   message.channel.send(exampleEmbed);
     
   }
-});
+  if(message.content.startsWith('!adminlog')) {
+    if(message.channel.id != "646099774797250575"){
+      message.delete()
+      return
+    }
+    if(!message.member._roles.includes(adminToken)){
+      message.reply("Na tohle nemas prava")
+      return
+    }
+    let bany
+    let kicky
+    //CHECK IF THERE ARE ANY BANNS OR KICKS
 
+    if(banns.length == 0){bany = "Zatim nebyli udeleny bany"}else{bany = banns}
+    if(kicks.length == 0){kicky = "Zatim nebyli udeleny zadne kicky"}else{kicky = kicks}
+    const exampleEmbed = new Discord.MessageEmbed()
+    .setColor('#0099ff')
+    .setTitle('**Admin LOG**')
+    .setThumbnail()
+    .addFields(
+        {name:"**Bany**", value: bany},
+        {name:"**Kicky**",value: kicky}
+    )
+    .setTimestamp()
+  
+  message.channel.send(exampleEmbed);
+    
+  }
+  if(message.content.startsWith('!serverlog')) {
+    if(message.channel.id != "646099774797250575"){
+      message.delete()
+      return
+    }
+    if(!message.member._roles.includes(adminToken)){
+      message.reply("Na tohle nemas prava")
+      return
+    }
+    //CHECK IF THERE ARE ANY BANNS OR KICKS
+    let log3
+    if(adlog.length == 0){log3 = "Zadna aktivita"}else{log3 = adlog}
+    const exampleEmbed = new Discord.MessageEmbed()
+    .setColor('#0099ff')
+    .setTitle('**Server LOG**')
+    .setThumbnail()
+    .addFields(
+        {name:"**LOG**",value: log3}
+    )
+    .setTimestamp()
+  message.channel.send(exampleEmbed);
+  }
+});
